requirements_gemini.txt
# Updated requirements.txt for Gemini-powered MCP Server

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Google Gemini AI
google-generativeai==0.3.2

# Database (keep your existing database)
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# HTTP client
aiohttp==3.9.1
httpx==0.25.2

# Environment and configuration
python-dotenv==1.0.0

# CORS and middleware
python-multipart==0.0.6

# Optional: For enhanced features
python-jose[cryptography]==3.3.0  # For JWT tokens
passlib[bcrypt]==1.7.4           # For password hashing 